+++ crates/nostr/CHANGELOG.md
@@ -30,6 +30,7 @@
 - Change the output and behavior of nip17::extract_relay_list and nip17::extract_owned_relay_list functions
 - Add `rand` feature (https://github.com/rust-nostr/nostr/pull/1167)
 - Add `os-rng` feature (https://github.com/rust-nostr/nostr/pull/1171)
+- Replace `TryIntoUrl` trait with `RelayUrlArg` enum (https://github.com/rust-nostr/nostr/pull/1217)
 
 ### Changed